services/requirements.txt
black
bottle
python-dotenv
mockredispy==2.9.3
redis
requests
pytest
webtest